requirements.txt
asn1crypto==0.24.0
cffi==1.12.3
cryptography==2.7
Django==2.2
django-cors-headers==2.5.3
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
jsonfield==2.0.2
jwt==0.6.1
Markdown==3.1.1
Pinax==0.9a2
pinax-eventlog==2.0.3
pycparser==2.19
PyJWT==1.7.1
PyPDF2==1.26.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
